+++ modules/elements/elements/hd-image-comparison/templates/template.php
@@ -1,6 +1,6 @@
 <?php
 
-/* Herzog Dupont for YOOtheme Pro Copyright (C) 2020–2022 Thomas Weidlich GNU GPL v3 */
+/* Herzog Dupont for YOOtheme Pro Copyright (C) 2020-2022 Thomas Weidlich GNU GPL v3 */
 
 $uniqid = uniqid('hd-');
 
@@ -20,6 +20,7 @@ $image_before = $this->el('image', [
     'loading' => $props['image_loading'] ? false : null,
     'width' => $props['image_width'],
     'height' => $props['image_height'],
+    'focal_point' => $props['image_before_focal_point'],
     'thumbnail' => true,
 ]);
 
@@ -37,6 +38,7 @@ $image_after = $this->el('image', [
     'loading' => $props['image_loading'] ? false : null,
     'width' => $props['image_width'],
     'height' => $props['image_height'],
+    'focal_point' => $props['image_after_focal_point'],
     'thumbnail' => true,
 ]);
 
@@ -64,6 +66,10 @@ $slider = $this->el('div', [
 
     'data-onmousemove' => [
         '{slider_onmousemove}',
+    ],
+
+    'data-margin-remove' => [
+        '{slider_margin_remove}',
     ],
 
 ]);