+++ modules/elements/elements/hd-toggle/element.php
@@ -1,6 +1,6 @@
 <?php
 
-/* Herzog Dupont for YOOtheme Pro Copyright (C) 2018–2022 Thomas Weidlich GNU GPL v3 */
+/* Herzog Dupont for YOOtheme Pro Copyright (C) 2018-2022 Thomas Weidlich GNU GPL v3 */
 
 namespace YOOtheme;
 
@@ -8,7 +8,7 @@ return [
     'transforms' => [
         'render' => function ($node) {
             // Don't render element if content fields are empty
-            return (Str::length($node->props['btn_label']) || $node->props['icon']) && (Str::length($node->props['content']) || Str::length($node->props['content2']));
+            return (Str::length($node->props['btn_label']) || $node->props['icon']) && (Str::length($node->props['content']) || Str::length($node->props['content2']) || Str::length($node->props['target']));
         },
     ],